Create README.md
Over the course of my career spanning 19 years, I have gained a wealth of experience in a variety of areas within the field of IT harnnesing skill of the whole SDLC. I have worked with a range of different clients to improve their overall IT operations. As an advocate of the Agile Project Methodology (Scrum), I always strive to deliver results quickly and efficiently, while ensuring that my work is of the highest quality. I am an energetic and dedicated IT professional who is always looking for new challenges and opportunities to learn and grow.